Various tweaks to get a working (probably?) 1.19.2 forge build.
I'm not familiar with Gradle, so some of this may be wrong or
unnecessary, but it was enough to get the mod loading and writing
to options.txt, so that's progress. I haven't found any UI for it,
though, and I'm not sure how to tell if it's working. But I wanted
to record what I've done anyways in case it's of interest.

Cheers and be well.